Modify chefspec file for change foodciritic error.

+++ site-cookbooks/backup_restore/recipes/configure_directory.rb
@@ -12,17 +12,16 @@ backup_model :directory do
 end
 
 # set proxy environment if use_proxy
-if node['backup_restore']['config']['use_proxy']
-  ruby_block 'set_proxy_env' do
-    block do
-      proxy_url = "http://#{node['backup_restore']['config']['proxy_host']}:#{node['backup_restore']['config']['proxy_port']}/"
+ruby_block 'set_proxy_env' do
+  block do
+    proxy_url = "http://#{node['backup_restore']['config']['proxy_host']}:#{node['backup_restore']['config']['proxy_port']}/"
 
-      cron_file = '/etc/cron.d/directory_backup'
-      file = Chef::Util::FileEdit.new(cron_file)
-      file.insert_line_after_match(/# Crontab for/, "https_proxy=#{proxy_url}")
-      file.insert_line_after_match(/# Crontab for/, "http_proxy=#{proxy_url}")
-      file.write_file
-      File.delete("#{cron_file}.old") if File.exist?("#{cron_file}.old")
-    end
+    cron_file = '/etc/cron.d/directory_backup'
+    file = Chef::Util::FileEdit.new(cron_file)
+    file.insert_line_after_match(/# Crontab for/, "https_proxy=#{proxy_url}")
+    file.insert_line_after_match(/# Crontab for/, "http_proxy=#{proxy_url}")
+    file.write_file
+    File.delete("#{cron_file}.old") if File.exist?("#{cron_file}.old")
   end
+  only_if { node['backup_restore']['config']['use_proxy'] }
 end
